pyproject.toml
[tool.poetry]
name = "vacuum-map-parser-xiaomi"
# The version is set by GH action on release
version = "0.0.0"
license = "Apache-2.0"
description = "Functionalities for Xiaomi vacuum map parsing"
readme = "README.md"
authors = ["Aron Kahrs <aronkahrs.u@gmail.com>","maksp86 <maksp@gsqd.ru>", "Alexander Vassilyevsky <tarh_@mail.ru>"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Environment :: Console",
    "Programming Language :: Python :: 3.11",
    "Topic :: Home Automation",
]
packages = [{ include = "vacuum_map_parser_xiaomi", from = "src" }]

[tool.poetry.urls]
"Homepage" = "https://github.com/aronkahrs-us/Python-package-vacuum-map-parser-xiaomi"
"Repository" = "https://github.com/aronkahrs-us/Python-package-vacuum-map-parser-xiaomi"
"Bug Tracker" = "https://github.com/aronkahrs-us/Python-package-vacuum-map-parser-xiaomi/issues"
"Changelog" = "https://github.com/aronkahrs-us/Python-package-vacuum-map-parser-xiaomi/releases"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "*"
pycryptodome = "*"

vacuum-map-parser-base = ">=0.1.5"

[tool.poetry.group.dev.dependencies]
black = "*"
mypy = "*"
ruff = "*"
isort = "*"
pylint = "*"
types-Pillow = "*"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
platform = "linux"
exclude = '''(?x)
    ( ^.*RobotMap_pb2\.py$
    | ^.*RobotMap_pb2\.pyi$
    | ^.*beautify_min\.py$
    )
'''

check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
no_implicit_reexport = true
strict_optional = true
warn_incomplete_stub = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pylint]
disable = [
    "C0103",
    "C0116",
    "R0902",
    "R0903",
    "R0912",
    "R0913",
    "R0914",
    "R0915",
    "W0640",
    "R0917",
]
max-line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"